project/requirements.txt
```
openai==0.27.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pymongo==4.3.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Pillow==9.5.0
pystray==0.19.4
psutil==5.9.5
wmi==1.5.1
pyautogui==0.9.54
```